+++ leather/scales/__init__.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+from leather.scales.annual import Annual
 from leather.scales.base import Scale
 from leather.scales.linear import Linear
 from leather.scales.ordinal import Ordinal